requirements.txt
numpy
opencv-python
matplotlib
scikit-image
scipy
mysql-connector-python
